android: move runtime state types to org.webrtc.audio, expose on factory
Relocate AudioProcessingImplementation, AudioProcessingComponentRuntimeState
and AudioProcessingRuntimeState out of PeerConnection.java into standalone
org.webrtc.audio classes, next to the AudioProcessingOptions/Mode/Result
types they belong with (mirroring the ObjC header relocation).

PeerConnectionFactory gains getAudioProcessingRuntimeState(), replacing the
removed PeerConnection accessor. The int-array JNI serializers move from
peer_connection.cc (where they were orphaned in an anonymous namespace) to
peer_connection_factory.cc, and the JNI entry point reads the state from
the factory.
