+++ sdk/android/BUILD.gn
@@ -329,6 +329,9 @@ if (is_android) {
       "api/org/webrtc/PeerConnection.java",
       "api/org/webrtc/PeerConnectionDependencies.java",
       "api/org/webrtc/PeerConnectionFactory.java",
+      "api/org/webrtc/audio/AudioProcessingComponentRuntimeState.java",
+      "api/org/webrtc/audio/AudioProcessingImplementation.java",
+      "api/org/webrtc/audio/AudioProcessingRuntimeState.java",
       "api/org/webrtc/RTCStats.java",
       "api/org/webrtc/RTCStatsCollectorCallback.java",
       "api/org/webrtc/RTCStatsReport.java",
