+++ sdk/android/src/jni/pc/peer_connection_factory.cc
@@ -15,11 +15,13 @@
 #include <unistd.h>
 
 #include <cstddef>
+#include <cstdint>
 #include <cstdio>
 #include <memory>
 #include <optional>
 #include <string>
 #include <utility>
+#include <vector>
 
 #include "absl/memory/memory.h"
 #include "api/audio/audio_device.h"
@@ -467,6 +469,79 @@ static jboolean JNI_PeerConnectionFactory_StartAecDump(
 static void JNI_PeerConnectionFactory_StopAecDump(JNIEnv* jni,
                                                   jlong native_factory) {
   PeerConnectionFactoryFromJava(native_factory)->StopAecDump();
+}
+
+// Serializers flattening AudioProcessingRuntimeState into the int array
+// layout expected by org.webrtc.audio.AudioProcessingRuntimeState.fromNative.
+static int32_t OptionalBoolToJava(std::optional<bool> value) {
+  return value.has_value() ? (*value ? 1 : 0) : -1;
+}
+
+static int32_t OptionalAudioProcessingModeToJava(std::optional<AudioProcessingMode> mode) {
+  return mode.has_value() ? static_cast<int32_t>(*mode) : -1;
+}
+
+static int32_t AudioProcessingImplementationToJava(AudioProcessingImplementation implementation) {
+  return static_cast<int32_t>(implementation);
+}
+
+static int32_t BuiltInAudioProcessingTopologyToJava(AudioDeviceModule::BuiltInAudioProcessingTopology topology) {
+  return static_cast<int32_t>(topology);
+}
+
+static void AppendAudioProcessingComponentRuntimeState(std::vector<int32_t> *values,
+                                                       const AudioProcessingComponentRuntimeState &state) {
+  values->push_back(OptionalBoolToJava(state.is_requested_enabled));
+  values->push_back(OptionalAudioProcessingModeToJava(state.requested_mode));
+  values->push_back(OptionalBoolToJava(state.is_resolved_software_enabled));
+  values->push_back(OptionalBoolToJava(state.is_software_enabled));
+  values->push_back(state.is_platform_available ? 1 : 0);
+  values->push_back(OptionalBoolToJava(state.is_platform_requested));
+  values->push_back(OptionalBoolToJava(state.is_platform_active));
+  values->push_back(AudioProcessingImplementationToJava(state.effective));
+}
+
+static void AppendBuiltInAudioProcessingComponentState(std::vector<int32_t> *values, bool available,
+                                                       std::optional<bool> requested,
+                                                       std::optional<bool> active) {
+  values->push_back(available ? 1 : 0);
+  values->push_back(OptionalBoolToJava(requested));
+  values->push_back(OptionalBoolToJava(active));
+}
+
+static void AppendBuiltInAudioProcessingState(std::vector<int32_t> *values,
+                                              const AudioDeviceModule::BuiltInAudioProcessingState &state) {
+  values->push_back(BuiltInAudioProcessingTopologyToJava(state.topology));
+  AppendBuiltInAudioProcessingComponentState(values, state.is_echo_cancellation_available,
+                                             state.is_echo_cancellation_requested, state.is_echo_cancellation_active);
+  AppendBuiltInAudioProcessingComponentState(values, state.is_noise_suppression_available,
+                                             state.is_noise_suppression_requested, state.is_noise_suppression_active);
+  AppendBuiltInAudioProcessingComponentState(values, state.is_auto_gain_control_available,
+                                             state.is_auto_gain_control_requested, state.is_auto_gain_control_active);
+}
+
+static std::vector<int32_t> AudioProcessingRuntimeStateToJavaValues(
+    const AudioProcessingRuntimeState &state) {
+  std::vector<int32_t> values;
+  values.reserve(47);
+  values.push_back(BuiltInAudioProcessingTopologyToJava(state.topology));
+  values.push_back(state.has_audio_processing_module ? 1 : 0);
+  values.push_back(state.has_audio_processing_config ? 1 : 0);
+  values.push_back(state.has_requested_audio_processing_options ? 1 : 0);
+  values.push_back(state.has_resolved_audio_processing_options ? 1 : 0);
+  AppendAudioProcessingComponentRuntimeState(&values, state.echo_cancellation);
+  AppendAudioProcessingComponentRuntimeState(&values, state.noise_suppression);
+  AppendAudioProcessingComponentRuntimeState(&values, state.auto_gain_control);
+  AppendAudioProcessingComponentRuntimeState(&values, state.high_pass_filter);
+  AppendBuiltInAudioProcessingState(&values, state.built_in);
+  return values;
+}
+
+static jni_zero::ScopedJavaLocalRef<jintArray>
+JNI_PeerConnectionFactory_GetAudioProcessingRuntimeState(JNIEnv* jni, jlong native_factory) {
+  return NativeToJavaIntArray(
+      jni, AudioProcessingRuntimeStateToJavaValues(
+               PeerConnectionFactoryFromJava(native_factory)->GetAudioProcessingRuntimeState()));
 }
 
 static jlong JNI_PeerConnectionFactory_CreatePeerConnection(
